+++ src/components/DishesSearchBar/DishesSearchBar.styled.js
@@ -1,6 +1,7 @@
 import styled from '@emotion/styled';
 
 export const SearchWrapper = styled.div`
+  padding: 24px;
   display: grid;
   grid-template-columns: repeat(4, 1fr);
   grid-column-gap: 12px;
