+++ src/pages/user/Dishes/DishesPage.styled.js
@@ -1,0 +1,9 @@
+import Typography from '@mui/material/Typography';
+
+import styled from '@emotion/styled';
+
+export const TypographyStyled = styled(Typography)`
+  margin-top: 16px;
+  text-align: center;
+  font-weight: 600;
+`;
